refactor: Separate operand-related classes
- Add classes: Integer, Real, Variable
