+++ Includes/Operands/Operand.hpp
@@ -22,57 +22,4 @@ public:
 	virtual std::string GetInfix(bool = false) const;
 };
 
-class Integer : public Operand
-{
-private:
-	int m_value;
-
-protected:
-	Integer(const Integer&);
-
-public:
-	Integer(int = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Real : public Operand
-{
-private:
-	double m_value;
-
-protected:
-	Real(const Real&);
-
-public:
-	Real(double = 0.0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Variable : public Operand
-{
-private:
-	std::string m_token;
-	VariableTable& m_varTable;
-
-protected:
-	Variable(const Variable&);
-
-public:
-	Variable(const std::string&, VariableTable&);
-
-	double GetValue() const;
-	void SetValue(const double&);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-
-};
-
 #endif
